Removed context from viewmodel to prevent leaking
